move assets to xlm/assets and merge files under horizon
since assets are stellar only assets, makes sense ot have them in a
stellar specific directory

also merge some files in xlm to make them more descriptive

also add rpc/complyadvantage to easyjson precommit script
