+++ platforms/opensolar/contract.go
@@ -6,7 +6,7 @@ import (
 	"log"
 	"time"
 
-	assets "github.com/YaleOpenLab/openx/assets"
+	assets "github.com/YaleOpenLab/openx/xlm/assets"
 	consts "github.com/YaleOpenLab/openx/consts"
 	database "github.com/YaleOpenLab/openx/database"
 	issuer "github.com/YaleOpenLab/openx/issuer"
