+++ platforms/ozones/contract.go
@@ -7,7 +7,7 @@ import (
 	"math"
 	"time"
 
-	assets "github.com/YaleOpenLab/openx/assets"
+	assets "github.com/YaleOpenLab/openx/xlm/assets"
 	consts "github.com/YaleOpenLab/openx/consts"
 	database "github.com/YaleOpenLab/openx/database"
 	issuer "github.com/YaleOpenLab/openx/issuer"
