+++ precommit.sh
@@ -2,5 +2,5 @@
 go fmt ./...
 go vet ./...
 ./misspell -w -q ./...
-./easyjson -all database/user.go database/recipient.go database/investor.go rpc/rpc.go platforms/opensolar/opensolar.go platforms/ozones/ozones.go rpc/users.go platforms/opensolar/entities.go
+./easyjson -all database/user.go database/recipient.go database/investor.go rpc/rpc.go platforms/opensolar/opensolar.go platforms/ozones/ozones.go rpc/users.go platforms/opensolar/entities.go rpc/complyadvantage.go
 ./staticcheck github.com/YaleOpenLab/openx/...
