+++ xlm/assets/assets.go
