+++ xlm/assets/assets_test.go
