+++ xlm/horizon.go
@@ -4,37 +4,84 @@ import (
 	"encoding/json"
 	"github.com/pkg/errors"
 	"io/ioutil"
+	"log"
 	"net/http"
 
-	"github.com/stellar/go/protocols/horizon"
+	clients "github.com/stellar/go/clients/horizon"
 	protocols "github.com/stellar/go/protocols/horizon"
 )
 
-/*
-type Account struct {
-    Links struct {
-        Self         hal.Link `json:"self"`
-        Transactions hal.Link `json:"transactions"`
-        Operations   hal.Link `json:"operations"`
-        Payments     hal.Link `json:"payments"`
-        Effects      hal.Link `json:"effects"`
-        Offers       hal.Link `json:"offers"`
-        Trades       hal.Link `json:"trades"`
-        Data         hal.Link `json:"data"`
-    }   `json:"_links"`
+// xlm is a set of functions that interface with stellar-core without needing the
+// horizon API that stellar provides, which is incomplete
 
-    HistoryAccount
-    Sequence             string            `json:"sequence"`
-    SubentryCount        int32             `json:"subentry_count"`
-    InflationDestination string            `json:"inflation_destination,omitempty"`
-    HomeDomain           string            `json:"home_domain,omitempty"`
-    Thresholds           AccountThresholds `json:"thresholds"`
-    Flags                AccountFlags      `json:"flags"`
-    Balances             []Balance         `json:"balances"`
-    Signers              []Signer          `json:"signers"`
-    Data                 map[string]string `json:"data"`
+// TestNetClient defines the horizon client to connect to
+var TestNetClient = &clients.Client{
+	// URL: "http://35.192.122.229:8080",
+	URL:  "https://horizon-testnet.stellar.org",
+	HTTP: http.DefaultClient,
 }
-*/
+
+// GetLedgerData gets the latest data from the ledger
+func GetLedgerData(blockNumber string) ([]byte, error) {
+	var err error
+	var data []byte
+	resp, err := http.Get(TestNetClient.URL + "/ledgers/" + blockNumber)
+	if err != nil || resp.Status != "200 OK" {
+		return data, errors.New("API Request did not succeed")
+	}
+	defer resp.Body.Close()
+	data, err = ioutil.ReadAll(resp.Body)
+	return data, err
+}
+
+// GetBlockHash gets the block hash corresponding to the passed block number
+func GetBlockHash(blockNumber string) (string, error) {
+	var err error
+	var hash string
+	b, err := GetLedgerData(blockNumber)
+	if err != nil {
+		return hash, errors.Wrap(err, "could not get updated ledger data")
+	}
+	var x protocols.Ledger
+	err = json.Unmarshal(b, &x)
+	hash = x.Hash
+	log.Printf("The block hash for block %d is: %s and the prev hash is %s", x.Sequence, hash, x.PrevHash)
+	return hash, err
+}
+
+// GetLatestBlockHash gets the lastest block hash
+func GetLatestBlockHash() (string, error) {
+	url := TestNetClient.URL + "/ledgers?cursor=now&order=desc&limit=1"
+	resp, err := http.Get(url)
+	if err != nil || resp.Status != "200 OK" {
+		return "", errors.New("API Request did not succeed")
+	}
+	defer resp.Body.Close()
+	data, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return "", errors.Wrap(err, "could not read response body")
+	}
+	// hacks below follow because of stellar's incomplete go sdk support
+	var x map[string]*json.RawMessage
+	err = json.Unmarshal(data, &x)
+	if err != nil {
+		return "", errors.Wrap(err, "could not unmarshal json")
+	}
+
+	var y map[string]*json.RawMessage
+	err = json.Unmarshal(*x["_embedded"], &y)
+	if err != nil {
+		return "", errors.Wrap(err, "could not unmarshal json")
+	}
+
+	var z []protocols.Ledger
+	err = json.Unmarshal(*y["records"], &z)
+	if err != nil {
+		return "", errors.Wrap(err, "could not unmarshal json")
+	}
+
+	return z[0].Hash, nil
+}
 
 // GetAccountData gets the account data
 func GetAccountData(a string) ([]byte, error) {
@@ -122,7 +169,7 @@ func GetAssetTrustLimit(publicKey string, assetName string) (string, error) {
 
 // GetAllBalances calls the stellar testnet API to get all the balances associated
 // with a certain account.
-func GetAllBalances(publicKey string) ([]horizon.Balance, error) {
+func GetAllBalances(publicKey string) ([]protocols.Balance, error) {
 	account, err := TestNetClient.LoadAccount(publicKey)
 	if err != nil {
 		return nil, errors.Wrap(err, "could not load account")
@@ -145,4 +192,34 @@ func HasStableCoin(PublicKey string) bool {
 		}
 	}
 	return false
+}
+
+// GetTransactionData gets tx data
+func GetTransactionData(txhash string) ([]byte, error) {
+	var err error
+	var data []byte
+	resp, err := http.Get(TestNetClient.URL + "/transactions/" + txhash)
+	if err != nil || resp.Status != "200 OK" {
+		// check here since if we don't, we need to check the body of the unmarshalled
+		// response to see if we have 0
+		return data, errors.New("API Request did not succeed")
+	}
+
+	defer resp.Body.Close()
+	data, err = ioutil.ReadAll(resp.Body)
+	return data, err
+}
+
+// GetTransactionHeight gets tx height
+func GetTransactionHeight(txhash string) (int, error) {
+	var err error
+	var txheight int
+
+	b, err := GetTransactionData(txhash)
+	if err != nil {
+		return txheight, errors.Wrap(err, "could not get transaction data")
+	}
+	var x protocols.Transaction
+	err = json.Unmarshal(b, &x)
+	return int(x.Ledger), err
 }
